portable_pty

Module serial

This module implements a serial port based tty. This is a bit different from the other implementations in that we cannot explicitly spawn a process into the serial connection, so we can only use a CommandBuilder::new_default_prog with the openpty method. On most (all?) systems, attempting to open multiple instances of the same serial port will fail.

Structs

SerialTty