portable_pty

Trait Child

pub trait Child: Debug + ChildKiller {
    // Required methods
    fn try_wait(&mut self) -> IoResult<Option<ExitStatus>>;
    fn wait(&mut self) -> IoResult<ExitStatus>;
    fn process_id(&self) -> Option<u32>;
}

Represents a child process spawned into the pty. This handle can be used to wait for or terminate that child process.

Required Methods

fn try_wait(&mut self) -> IoResult<Option<ExitStatus>>

Poll the child to see if it has completed. Does not block. Returns None if the child has not yet terminated, else returns its exit status.

fn wait(&mut self) -> IoResult<ExitStatus>

Blocks execution until the child process has completed, yielding its exit status.

fn process_id(&self) -> Option<u32>

Returns the process identifier of the child process, if applicable