portable_pty

Trait PtySystem

pub trait PtySystem: Downcast {
    // Required method
    fn openpty(&self, size: PtySize) -> Result<PtyPair>;
}

The PtySystem trait allows an application to work with multiple possible Pty implementations at runtime. This is important on Windows systems which have a variety of implementations.

Required Methods

fn openpty(&self, size: PtySize) -> Result<PtyPair>

Create a new Pty instance with the window size set to the specified dimensions. Returns a (master, slave) Pty pair. The master side is used to drive the slave side.

Implementations

impl dyn PtySystem

pub fn is<__T: PtySystem>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: PtySystem>(self: Box<Self>) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: PtySystem>(self: Rc<Self>) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: PtySystem>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: PtySystem>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.