portable_pty

Module unix

Working with pseudo-terminals

Structs

UnixPtySystem

Functions

close_random_fds

On Big Sur, Cocoa leaks various file descriptors to child processes, so we need to make a pass through the open descriptors beyond just the stdio descriptors and close them all out. This is approximately equivalent to the darwin posix_spawnattr_setflags option POSIX_SPAWN_CLOEXEC_DEFAULT which is used as a bit of a cheat on macOS. On Linux, gnome/mutter leak shell extension fds to wezterm too, so we also need to make an effort to clean up the mess.

Type Aliases

RawFd

Raw file descriptors.