portable_pty

Trait ChildKiller

pub trait ChildKiller: Debug {
    // Required methods
    fn kill(&mut self) -> IoResult<()>;
    fn clone_killer(&self) -> Box<dyn ChildKiller + Send + Sync>;
}

Represents the ability to signal a Child to terminate

Required Methods

fn kill(&mut self) -> IoResult<()>

Terminate the child process

fn clone_killer(&self) -> Box<dyn ChildKiller + Send + Sync>

Clone an object that can be split out from the Child in order to send it signals independently from a thread that may be blocked in .wait.